.references/validation/requirements.txt
# Python dependencies for ephemeris validation
# Install with: pip install -r requirements.txt

pyswisseph>=2.10.0

